requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
cryptography==41.0.7
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
ecdsa==0.18.0
executing==2.0.1
fastapi==0.108.0
fireworks-ai==0.11.0
frozenlist==1.4.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
httpx-sse==0.4.0
idna==3.6
importlib-metadata==6.11.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
langchain==0.0.335
langchain-community==0.0.9
langchain-core==0.1.7
langchainplus-sdk==0.0.20
langsmith==0.0.77
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numexpr==2.8.7
numpy==1.26.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.1.3
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==1.4.50
stack-data==0.6.3
starlette==0.32.0.post1
streamlit==1.24.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.1.0
validators==0.22.0
wcwidth==0.2.12
yarl==1.9.4
zipp==3.17.0
